refine unified trainer API
