+++ doc/code-docs/source/example/7B_demo.rst
@@ -167,7 +167,7 @@
 
     srun -p internllm -N 1 -n 8 --ntasks-per-node=8 --gpus-per-task=1 python train.py --config ./configs/7B_sft.py
 
-其中，train.py文件的编写参考：[训练代码示例](../../../../examples/huggingface_trainer/train.py)
+其中，train.py文件的编写参考：[训练代码示例](../../../../train.py)
 
 训练结果
 ----------------
