+++ internlm/data/streaming/__init__.py
@@ -1,7 +1,7 @@
 from .batch_sampler import StreamingStaticBatchSampler
 from .collaters import nopack_collate_fn, pack_collate_fn
 from .dataset import HuggingFacePackedDataset, HuggingFaceStreamingDataset
-from .utils import naive_hf_resume
+from .utils import hf_simple_resume
 
 __all__ = [
     "StreamingStaticBatchSampler",
@@ -9,5 +9,5 @@ __all__ = [
     "pack_collate_fn",
     "HuggingFaceStreamingDataset",
     "HuggingFacePackedDataset",
-    "naive_hf_resume",
+    "hf_simple_resume",
 ]
