+++ internlm/monitor/__init__.py
@@ -1,5 +1,4 @@
-from .internevo_monitor import internevo_monitor
-from .monitor import initialize_monitor_manager, send_alert_message
+from .monitor import initialize_monitor_manager, internevo_monitor, send_alert_message
 from .utils import set_env_var
 
 __all__ = [
